+++ NEWS.md
@@ -1,3 +1,9 @@
+# mlr3pipelines 0.6.0
+
+* Compatibility with new `bbotk` release.
+* Added marshaling support to `GraphLearner`
+* Support internal tuning and validation
+
 # mlr3pipelines 0.5.2
 
 * Added new `ppl("convert_types")`.